// packages/datadog-api-client-v1/models/DashboardTemplateVariable.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Template variable.
*/
export declare class DashboardTemplateVariable {
    /**
    * The default value for the template variable on dashboard load.
    */
    '_default'?: string;
    /**
    * The name of the variable.
    */
    'name': string;
    /**
    * The tag prefix associated with the variable. Only tags with this prefix appear in the variable drop-down.
    */
    'prefix'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): DashboardTemplateVariable;
    static serialize(data: DashboardTemplateVariable): {
        [key: string]: any;
    };
    constructor();
}